backend/requirements.txt
fastapi
uvicorn
gunicorn
aiofiles
pydantic
firebase-admin
bcrypt
google-cloud-translate
google-cloud-bigquery
google-cloud-aiplatform
google-cloud-storage
python-dotenv
python-multipart
openai

# 其他依賴
